feature : "ci 테스트 오류 수정"
홈 탑 바 파일에 알림 버튼 통합, getx put으로 받게 수정, 몇몇 stf를 stl로 변환

Resolves: #12
