+++ lib/presentation/home/widgets/home_top_bar.dart
@@ -1,26 +1,14 @@
 import 'package:flutter/material.dart';
 import 'package:ondo/core/design_system/app_colors.dart';
 import 'package:ondo/core/design_system/app_text_styles.dart';
-import 'package:ondo/presentation/home/widgets/home_alert_button.dart';
+import 'package:get/get.dart';
+import 'package:ondo/presentation/home/controllers/home_alert_controller.dart';
 
-class HomeTopBar extends StatefulWidget {
+class HomeTopBar extends StatelessWidget {
 
-  const HomeTopBar({super.key});
+  HomeTopBar({super.key});
 
-  @override
-  State<HomeTopBar> createState() => _HomeTopBarState();
-}
-
-class _HomeTopBarState extends State<HomeTopBar> {
-  late TextEditingController _searchController;
-
-
-  @override
-  void initState() {
-    _searchController = TextEditingController();
-    super.initState();
-  }
-
+  late TextEditingController _searchController = TextEditingController();
 
   final height = 44.0;
 
@@ -60,6 +48,69 @@ class _HomeTopBarState extends State<HomeTopBar> {
             child: HomeAlertButton(size: height,),
           ),
         ],
+      ),
+    );
+  }
+}
+
+class HomeAlertButton extends StatelessWidget {
+  final _size;
+
+  HomeAlertButton({super.key, required dynamic size}) : _size = size;
+
+  final alertController = Get.put(HomeAlertController());
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      width: _size,
+      height: _size,
+      alignment: Alignment.center,
+      decoration: BoxDecoration(
+        color: AppColors.gray20,
+        borderRadius: BorderRadius.circular(8),
+      ),
+      child: SizedBox.expand(
+        child: Stack(
+          children: [
+            Align(
+              alignment: Alignment.center,
+              child: Obx(
+                    () => Icon(
+                  Icons.add_alert,
+                  size: 24,
+                  color: alertController.enable.value
+                      ? AppColors.primaryDark
+                      : AppColors.gray40,
+                ),
+              ),
+            ),
+
+            Positioned(
+              top: 9,
+              right: 5,
+              child: Obx(
+                    () => Container(
+                  padding: EdgeInsets.all(2),
+                  alignment: Alignment.center,
+                  decoration: BoxDecoration(
+                    color: Colors.red,
+                    borderRadius: BorderRadius.circular(6),
+                  ),
+                  child: Text(
+                    "${alertController.totals.value}",
+                    style: TextStyle(
+                      color: Colors.white,
+                      fontSize: 8,
+                      fontWeight: FontWeight.bold,
+                    ),
+                    textAlign: TextAlign.center,
+                  ),
+                ),
+              ),
+            ),
+          ],
+        ),
       ),
     );
   }
